max_instances for DatasetReader

+++ allennlp/tests/data/dataset_readers/dataset_reader_test.py
@@ -1,22 +1,23 @@
 import os
 import shutil
 
+import pytest
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.dataset_readers import SnliReader
 from allennlp.data.dataset_readers.dataset_reader import _LazyInstances
 
 
-class DatasetReaderTest(AllenNlpTestCase):
-    def setUp(self):
-        super().setUp()
-        self.cache_directory = str(self.FIXTURES_ROOT / "data_cache" / "with_prefix")
+class TestDatasetReader:
+    cache_directory = str(AllenNlpTestCase.FIXTURES_ROOT / "data_cache" / "with_prefix")
 
-    def tearDown(self):
-        super().tearDown()
-        shutil.rmtree(self.cache_directory)
+    @pytest.yield_fixture(autouse=True)
+    def cache_directory_fixture(self):
+        yield self.cache_directory
+        if os.path.exists(self.cache_directory):
+            shutil.rmtree(self.cache_directory)
 
     def test_read_creates_cache_file_when_not_present(self):
-        snli_file = self.FIXTURES_ROOT / "data" / "snli.jsonl"
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
         reader = SnliReader(cache_directory=self.cache_directory)
         cache_file = reader._get_cache_location_for_file_path(snli_file)
         assert not os.path.exists(cache_file)
@@ -24,7 +25,7 @@ class DatasetReaderTest(AllenNlpTestCase):
         assert os.path.exists(cache_file)
 
     def test_read_uses_existing_cache_file_when_present(self):
-        snli_file = self.FIXTURES_ROOT / "data" / "snli.jsonl"
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
         snli_copy_file = str(snli_file) + ".copy"
         shutil.copyfile(snli_file, snli_copy_file)
         reader = SnliReader(cache_directory=self.cache_directory)
@@ -40,7 +41,7 @@ class DatasetReaderTest(AllenNlpTestCase):
             assert instance.fields == cached_instance.fields
 
     def test_read_only_creates_cache_file_once(self):
-        snli_file = self.FIXTURES_ROOT / "data" / "snli.jsonl"
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
         reader = SnliReader(cache_directory=self.cache_directory)
         cache_file = reader._get_cache_location_for_file_path(snli_file)
 
@@ -61,7 +62,7 @@ class DatasetReaderTest(AllenNlpTestCase):
         assert cache_contents == final_cache_contents
 
     def test_caching_works_with_lazy_reading(self):
-        snli_file = self.FIXTURES_ROOT / "data" / "snli.jsonl"
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
         snli_copy_file = str(snli_file) + ".copy"
         shutil.copyfile(snli_file, snli_copy_file)
         reader = SnliReader(lazy=True, cache_directory=self.cache_directory)
@@ -98,3 +99,27 @@ class DatasetReaderTest(AllenNlpTestCase):
         assert len(first_pass_instances) == len(cached_instances)
         for instance, cached_instance in zip(first_pass_instances, cached_instances):
             assert instance.fields == cached_instance.fields
+
+    @pytest.mark.parametrize("lazy", (True, False))
+    def test_max_instances(self, lazy):
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
+        reader = SnliReader(max_instances=2, lazy=lazy)
+        instances = reader.read(snli_file)
+        instance_count = sum(1 for _ in instances)
+        assert instance_count == 2
+
+    @pytest.mark.parametrize("lazy", (True, False))
+    def test_cached_max_instances(self, lazy):
+        snli_file = AllenNlpTestCase.FIXTURES_ROOT / "data" / "snli.jsonl"
+
+        # The first read will create the cache if it's not there already.
+        reader = SnliReader(cache_directory=self.cache_directory, lazy=lazy)
+        instances = reader.read(snli_file)
+        instance_count = sum(1 for _ in instances)
+        assert instance_count > 2
+
+        # The second read should only return two instances, even though it's from the cache.
+        reader = SnliReader(cache_directory=self.cache_directory, max_instances=2, lazy=lazy)
+        instances = reader.read(snli_file)
+        instance_count = sum(1 for _ in instances)
+        assert instance_count == 2
